# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: analyses
    description: Trading analysis operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /analyze:
    post:
      operationId: createAnalysis
      tags: [analyses]
      summary: Start a new trading analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnalysisBody"
      responses:
        "201":
          description: Analysis created and job started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Python agent service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /analyses:
    get:
      operationId: listAnalyses
      tags: [analyses]
      summary: List all analyses
      responses:
        "200":
          description: List of analyses ordered by most recent first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Analysis"

  /analyses/{id}:
    get:
      operationId: getAnalysis
      tags: [analyses]
      summary: Get a specific analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
        "404":
          description: Analysis not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /analyses/{id}/stream:
    get:
      operationId: streamAnalysis
      tags: [analyses]
      summary: Stream real-time agent progress for an analysis
      description: |
        Server-Sent Events (SSE) stream. Emits JSON event objects with a `type` field.
        Event types: `started`, `agent_update`, `agent_report`, `completed`, `error`, `done`.
        Connect using the browser's `EventSource` API.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: SSE stream of agent progress events
          content:
            text/event-stream:
              schema:
                type: string
                description: Server-Sent Events stream
        "404":
          description: Analysis not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    CreateAnalysisBody:
      type: object
      properties:
        ticker:
          type: string
          description: Stock ticker symbol (e.g. NVDA, AAPL)
        date:
          type: string
          description: Analysis date in YYYY-MM-DD format
        model:
          type: string
          description: OpenRouter model identifier
          default: minimax/minimax-m2.5
        maxDebateRounds:
          type: integer
          description: Number of debate rounds between bull/bear researchers
          default: 1
      required:
        - ticker
        - date

    Analysis:
      type: object
      properties:
        id:
          type: integer
        ticker:
          type: string
        date:
          type: string
        model:
          type: string
        status:
          type: string
          enum: [pending, running, completed, error]
        decision:
          type: ["string", "null"]
        reasoning:
          type: ["string", "null"]
        jobId:
          type: string
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - ticker
        - date
        - model
        - status
        - jobId
        - createdAt
        - updatedAt

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error